feat(PlayerStandingTest): flop quads test
