+++ src/test/scala/EquityTest/StreetTest.scala
@@ -11,7 +11,7 @@ import poker.Hand._
 import poker.OrderInstances.{handOrder, handOrdering}
 import pokerData.BoardGenerators._
 
-object StreetTest extends Properties("BoardState Tests") {
+object StreetTest extends Properties("Street Tests") {
 
   // genFlop, genTurn, genRiver
   // genPlayer,
@@ -44,29 +44,6 @@ object StreetTest extends Properties("BoardState Tests") {
     allDealtCards.distinct ?= allDealtCards
 
   }
-  property("Preflop: If the winner is a pair, no more than two players can be winning ") = forAll(genPreflopBoard) {
-    preFlop =>
-      val winnerList = PlayerStanding.winnerList(preFlop)
-
-      val winningRank = winnerList.get.map(_._3).fold(HighCard(Two, List.empty[Card])) { case (s: Hand, v: Hand) =>
-        if (handOrder.compare(s, v) > 0) s else v
-      }
-      if ((winningRank.score == 2) && (winnerList.get.size > 2)) false
-      else if (winningRank.score > 2) false
-      else true
-  }
-
-  property("Preflop: If the winner is a HighCard, no more than 4 players can be winning ") = forAll(genPreflopBoard) {
-    preFlop =>
-      val winnerList = PlayerStanding.winnerList(preFlop)
-
-      val winningRank = winnerList.get.map(_._3).fold(HighCard(Two, List.empty[Card])) { case (s: Hand, v: Hand) =>
-        if (handOrder.compare(s, v) > 0) s else v
-      }
-      if ((winningRank.score == 1) && (winnerList.get.size > 4)) false
-      else if (winningRank.score > 2) false
-      else true
-  }
 
   //  property("preflop deck size equals 52 minus the players cards and maintains uniqueness") = forAll(genPreflopBoard, genNumberOfPlayers) {
 //    (preFlopBoard, numPlayers) =>
@@ -79,15 +56,5 @@ object StreetTest extends Properties("BoardState Tests") {
 //      playerCards.distinct.size == playerCards.size &&
 //      preFlopBoard.deck.cards.distinct.size == preFlopBoard.deck.size
 //  }
-
-// TODO This is not a ShowDown operation, I'm asking what the current hand street is!!!
-  property("Flop: more than one player can NOT have quads") = forAll(genFlopBoard) { flop =>
-    ShowDown(flop.allHands).count(p =>
-      p match {
-        case _: FourOfAKind => true
-        case _              => false
-      }
-    ) <= 1
-  }
 
 }
